pyproject.toml
[project]
name = "git-reading"
version = "0.1.0"
description = "Index reading notes from markdown files using git history"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",  # Beautiful CLI output and logging
    "python-dotenv>=1.0.0",  # Load environment variables from .env file
    "requests>=2.31.0",  # HTTP library for API clients
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "psycopg[binary,pool]>=3.1.0",  # PostgreSQL driver for testing
]
test-full = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "psycopg[binary,pool]>=3.1.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]
streamlit = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "watchdog>=6.0.0",
]
search = [
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
]
postgresql = [
    "psycopg[binary,pool]>=3.1.0",  # PostgreSQL driver with connection pooling
]
api = [
    "fastapi>=0.104.0",
    "strawberry-graphql[fastapi]>=0.217.0",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
extract = "extract.cli:main"
normalize = "normalize_source.cli:main"
search = "query.cli:main"
load-db = "load.cli:main"
enrich-db = "enrich.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/api",
    "src/common",
    "src/extract",
    "src/enrich",
    "src/load",
    "src/normalize_source",
    "src/query",
    "src/transform",
]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["api", "common", "extract", "enrich", "load", "normalize_source", "query", "transform", "search"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "F401", "F811", "E402"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/streamlit_app/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
skip_empty = true
precision = 2
